🚚 Move: `TagPicker.svelte` and `blog/Tag.svelte` to `$lib/components/tag-picker/` [#86]
