+++ src/routes/blog/+page.svelte
@@ -4,7 +4,7 @@
 	import Title from '$lib/components/Title.svelte';
 	import FeedButton from '$lib/components/FeedButton.svelte';
 	import ShareButton from '$lib/components/ShareButton.svelte';
-	import TagPicker from './TagPicker.svelte';
+	import TagPicker from '$lib/components/tag-picker/TagPicker.svelte';
 	import ArticleCard from '$lib/components/ArticleCard.svelte';
 
 	import type { PageData } from './$types';
